Populating button and LED header files
Creating the button and LED header files. Populated the button header file with what I believe to be all of the required information for the final button setup. Need clarification on the LED setup before doing the same for the LED header file.
Added the lib_deps to the platformio ini file